Refactor: from "CouponSummary" to "CouponDetails", add "details" package
