+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/coupons/details/CouponDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.coupons
+package com.woocommerce.android.ui.coupons.details
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,24 +7,24 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.viewModels
 import com.woocommerce.android.R
-import com.woocommerce.android.databinding.FragmentCouponSummaryBinding
+import com.woocommerce.android.databinding.FragmentCouponDetailsBinding
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class CouponSummaryFragment : BaseFragment(R.layout.fragment_coupon_summary) {
-    private var _binding: FragmentCouponSummaryBinding? = null
+class CouponDetailsFragment : BaseFragment(R.layout.fragment_coupon_details) {
+    private var _binding: FragmentCouponDetailsBinding? = null
     private val binding get() = _binding!!
 
-    private val viewModel: CouponSummaryViewModel by viewModels()
+    private val viewModel: CouponDetailsViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentCouponSummaryBinding.inflate(inflater, container, false)
+        _binding = FragmentCouponDetailsBinding.inflate(inflater, container, false)
 
         val view = binding.root
         binding.couponsComposeView.apply {
@@ -32,7 +32,7 @@ class CouponSummaryFragment : BaseFragment(R.layout.fragment_coupon_summary) {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
             setContent {
                 WooThemeWithBackground {
-                    CouponSummaryScreen(viewModel)
+                    CouponDetailsScreen(viewModel)
                 }
             }
         }
