+++ WooCommerce/src/main/res/layout/fragment_coupon_details.xml
